+++ src/chattool/__init__.py
@@ -4,7 +4,7 @@ import importlib
 
 __author__ = """Rex Wang"""
 __email__ = "1073853456@qq.com"
-__version__ = "7.0.3"
+__version__ = "7.2.0"
 
 from dotenv import load_dotenv